+++ doc/release-notes.md
@@ -5,3 +5,4 @@ Bitcoin ABC version 0.16.2 is now available from:
 This release includes the following features and fixes:
 
  - Remove the newdaaactivationtime configuration.
+ - Do not use the NODE_BITCOIN_CASH service bit for preferencial peering anymore.